Restructure the returned entries, add Net::FTP::List::Entry
Make a number of changes so that the parsers are classes and
there is no object allocation when going through the parser set.
Every parser just has class methods and there will be only one
parser object per process. When parsing fails, return +false+
instead of raising since exceptions might be expensive and they
will occur on pretty much every iteration. Add an optional
parameter to raise if nothing can be parsed so that more strict
situations are handled gracefully.